Consolidate filter functions into filter.py
